[BE] feat : 완료한 미션 삭제 불가하게 수정

+++ backend/src/main/java/com/ssafy/hugging/mission/service/MissionService.java
@@ -65,6 +65,8 @@ public class MissionService {
 		ProceedingMission proceedingMission = proceedingMissionRepository.findProceedingMissionByMission_IdAndMember_Id(
 				missionId, memberId)
 			.orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_PROCEEDING_MISSION_ERROR_MESSAGE));
+		if(proceedingMission.getStatus() == Status.COMPLETE)
+			throw new IllegalArgumentException(DELETE_COMPLETE_MISSION_ERROR_MESSAGE);
 		proceedingMissionRepository.delete(proceedingMission);
 	}
